refactor(shared): extract model feature helpers into shared.model_features
Model knowledge feature checks are shared by plugins that decide whether note or knowledge tools should be available. Move them to a single helper module.
