+++ src/owui_ext/tools/sub_agent.py
@@ -38,6 +38,10 @@ from pydantic import BaseModel, Field
 
 from owui_ext.shared.async_utils import maybe_await
 from owui_ext.shared.builtin_tools import BUILTIN_TOOL_CATEGORIES, VALVE_TO_CATEGORY
+from owui_ext.shared.model_features import (
+    model_has_note_knowledge,
+    model_knowledge_tools_enabled,
+)
 from owui_ext.shared.notifications import emit_notification
 from owui_ext.shared.prompt_utils import (
     _append_tool_server_prompts,
@@ -170,26 +174,6 @@ def normalize_parallel_sub_agent_tasks(tasks: Any) -> tuple[Optional[list[dict[s
         validated_tasks.append({"description": description, "prompt": prompt})
 
     return validated_tasks, None
-
-
-def model_has_note_knowledge(model: Optional[dict]) -> bool:
-    """Return True if the current model has note-type attached knowledge."""
-    if not isinstance(model, dict):
-        return False
-    knowledge_items = (model.get("info", {}).get("meta", {}).get("knowledge") or [])
-    if not isinstance(knowledge_items, list):
-        return False
-    return any(item.get("type") == "note" for item in knowledge_items if isinstance(item, dict))
-
-
-def model_knowledge_tools_enabled(model: Optional[dict]) -> bool:
-    """Return True if model-level builtin knowledge tools are enabled."""
-    if not isinstance(model, dict):
-        return True
-    builtin_tools = model.get("info", {}).get("meta", {}).get("builtinTools", {})
-    if not isinstance(builtin_tools, dict):
-        return True
-    return bool(builtin_tools.get("knowledge", True))
 
 
 async def resolve_terminal_id_for_sub_agent(
